+++ src/App.js
@@ -1,16 +1,16 @@
 import React from 'react';
-import Body from './components/Body';
-import Header from './components/Header'
+import Body from './components/body/Body';
+import Header from './components/header/Header';
 
 class App extends React.Component {
-  render() {
-    return(
-      <>
-        <Header />
-        <Body />
-      </>
-    )
-  }
+    render() {
+        return (
+            <>
+                <Header />
+                <Body />
+            </>
+        );
+    }
 }
 
 export default App;
