+++ src/components/body/Body.css
@@ -1,0 +1,5 @@
+.divBody {
+    position: absolute;
+    width: 100%;
+    background-color: rgb(87, 87, 87);
+}
