+++ src/components/body/cards/CardList.css
@@ -1,0 +1,13 @@
+.labelDel {
+    display: flex;
+    width: 170px;
+    margin-left: 10px;
+    cursor: pointer;
+}
+h2 {
+    color: white;
+}
+.divCards {
+    display: flex;
+    flex-wrap: wrap;
+}
