+++ src/index.css
@@ -1,74 +1,13 @@
 body {
-  margin: 0;
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
-    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
-    sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
+    margin: 0;
+    font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto',
+        'Oxygen', 'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans',
+        'Helvetica Neue', sans-serif;
+    -webkit-font-smoothing: antialiased;
+    -moz-osx-font-smoothing: grayscale;
 }
 
 code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
-    monospace;
-}
-
-.divBody {
-  position: absolute;
-  width: 100%;
-  background-color: rgb(87, 87, 87);
-}
-.divCards {
-  display: flex;
-  flex-wrap: wrap;
-}
-.divCard {
-  width: 400px;
-  height: 500px;
-  margin: 10px;
-  margin-left: auto;
-  margin-right: auto;
-  background-color: rgba(0, 0, 0, 0.39);
-  color: white;
-  border: 2px ridge white;
-}
-.new {
-  background-color: rgba(201, 25, 25, 0.39); 
-  border: 2px solid black;
-}
-.cardHeader {
-  width: 95%;
-  height: 50px;
-  margin: 10px auto;
-  display: flex;
-}
-.input {
-  background-color: rgba(138, 138, 138, 0.493);
-  font-size: 19px;
-  font-weight: bolder;
-  color: rgb(255, 255, 255);
-  padding: 10px;
-}
-.checkbox {
-  transform: scale(2.0);
-}
-.titleCard {
-  width: 70%;
-  margin: auto 10px;
-  white-space: nowrap; 
-  overflow: hidden; 
-  text-overflow: ellipsis;
-}
-textarea {
-  width: 95%;
-  height: 350px;
-  margin: 0 auto;
-  padding: 10px;
-  outline: none;
-  border: 2px inset rgb(134, 134, 134);
-  background: none;
-  color: white;
-  resize: none;
-}
-.bodyText {
-  word-break: break-all;
+    font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
+        monospace;
 }
